separate URLs for interactive chart
