+++ ui/src/app/app-routing.module.ts
@@ -6,13 +6,14 @@ import { ChartDisplayComponent } from './chart-display/chart-display.component';
 import { ChartEditComponent } from './chart-edit/chart-edit.component';
 import { ChartSelectComponent } from './chart-select/chart-select.component';
 import { CreateBasicChartComponent } from './create-chart-basic/create-chart-basic.component';
-import { CreateInteractiveChartComponent } from './create-chart-interactive/create-chart-interactive.component';
 import { SeriesSelectComponent } from './series-select/series-select.component';
 import { SongDisplayComponent } from './song-display/song-display.component';
 import { SongComponent } from './song/song.component';
 import { TotalsComponent } from './totals/totals.component';
 import { HomeComponent } from './home/home.component';
 import { CreateChartComponent } from './create-chart/create-chart.component';
+import { EnterSongsComponent } from './create-chart-interactive/enter-songs/enter-songs.component';
+import { RankSongsComponent } from './create-chart-interactive/rank-songs/rank-songs.component';
 
 const routes: Routes = [
   // Can't have 'children' unless we want the parent displayed also!
@@ -27,8 +28,13 @@ const routes: Routes = [
      },
      {
       path: 'interactive',
-      component: CreateInteractiveChartComponent,
-     }]
+      component: EnterSongsComponent
+     },
+     {
+      path: 'interactive/:session/rank',
+      component: RankSongsComponent
+     }
+    ]
   },
   { path: 'series/:series/chart/:name', component: ChartDisplayComponent },
   { path: 'series/:series/chart/:name/edit', component: ChartEditComponent },
