--- ui/src/app/create-chart-interactive/create-chart-interactive.component.html
@@ -1,2 +1,0 @@
-<app-enter-songs *ngIf="screenToDisplay == 'enter'" [seriesName]="seriesName" (sessionIdEvent)="moveToRanking($event)"></app-enter-songs>
-<app-rank-songs *ngIf="screenToDisplay == 'rank'" [sessionId]="sessionId"></app-rank-songs>
