+++ ui/src/app/create-chart-interactive/enter-songs/enter-songs.component.spec.ts
@@ -2,14 +2,25 @@ import { ComponentFixture, TestBed } from '@angular/core/testing';
 
 import { EnterSongsComponent } from './enter-songs.component';
 import { HttpClientModule } from '@angular/common/http';
+import { of } from 'rxjs';
+import { ActivatedRoute } from '@angular/router';
 
 
 describe('EnterSongsComponent', () => {
   let component: EnterSongsComponent;
   let fixture: ComponentFixture<EnterSongsComponent>;
 
+  const mockActivatedRoute = {
+    parent: {
+      params: of({ series: 'test' }),
+    }
+  };
+
   beforeEach(async () => {
-    await TestBed.configureTestingModule({      
+    await TestBed.configureTestingModule({     
+      providers: [
+        {provide: ActivatedRoute, useValue: mockActivatedRoute},
+      ], 
       imports: [HttpClientModule],
       declarations: [ EnterSongsComponent ]
     })
