+++ ui/src/app/create-chart-interactive/enter-songs/enter-songs.component.ts
@@ -2,6 +2,7 @@ import { Component, Input, Output } from '@angular/core';
 import { AbstractControl, FormControl, FormGroup, ValidatorFn } from '@angular/forms';
 import { ChartService } from 'src/app/services/chart.service';
 import { EventEmitter } from '@angular/core';
+import { ActivatedRoute, Router } from '@angular/router';
 
 @Component({
   selector: 'app-enter-songs',
@@ -9,8 +10,6 @@ import { EventEmitter } from '@angular/core';
   styleUrls: ['../../styles/common-styles.css', './enter-songs.component.css']
 })
 export class EnterSongsComponent {
-    @Input() seriesName: string
-    @Output() sessionIdEvent = new EventEmitter();
 
     public chartService: ChartService
     public useDateAsTitle = false
@@ -24,10 +23,23 @@ export class EnterSongsComponent {
       revealOrder: new FormControl<'random' | 'inOrder'>('random')
     });
 
+    private router: Router
+    private activatedRoute: ActivatedRoute
 
-    public constructor(chartService: ChartService) {
+    public seriesName: string
+
+    public constructor(chartService: ChartService, router: Router, activatedRoute: ActivatedRoute) {
       this.chartService = chartService
+      this.router = router
+      this.activatedRoute = activatedRoute
     }
+
+    ngOnInit(): void {
+      this.activatedRoute.parent.params.subscribe((params => {
+        if (params.series) {
+          this.seriesName = params.series
+        }
+    }))}
 
     // If we're using reactive forms we are unable to use the [disabled] attribute
     // therefore we need to explicitly disable/enable the textbox
@@ -58,7 +70,7 @@ export class EnterSongsComponent {
 
     public onSubmit() {
       this.chartService.initiateInteractiveChartCreation(this.seriesName, this.chartForm.value).subscribe(res => {
-        this.sessionIdEvent.emit(res.sessionId)
+        this.router.navigate([res.sessionId, 'rank'], {relativeTo: this.activatedRoute})
       })
     }
