+++ ui/src/app/create-chart-interactive/rank-songs/rank-songs.component.html
@@ -1,1 +1,1 @@
-<p>rank-songs works!</p>
+<p>{{sessionId}}</p>
